+++ src/wembed_core/controllers/dl_document_controller.py
@@ -6,7 +6,7 @@ from datetime import datetime, timezone
 from typing import List, Optional
 
 from wembed_core.database import DatabaseService
-from wembed_core.models.dl_documents import DLDocuments
+from wembed_core.models.dl_doc.dl_documents import DLDocuments
 from wembed_core.schemas.dl_schemas import DLDocumentSchema
