+++ src/wembed_core/controllers/dl_input_controller.py
@@ -5,7 +5,7 @@ wembed_core/controllers/dl_input_controller.py
 from typing import List, Optional
 
 from wembed_core.database import DatabaseService
-from wembed_core.models.dl_inputs import DLInputs
+from wembed_core.models.dl_doc.dl_inputs import DLInputs
 from wembed_core.schemas.dl_schemas import DLInputSchema
