+++ src/wembed_core/models/__init__.py
@@ -1,4 +1,5 @@
 from .code_chunker import *  # noqa: F403, F401
+from .dl_doc import *  # noqa: F403, F401
 from .indexed_files import IndexedFiles  # noqa: F401
 from .indexed_obsidian_vaults import IndexedObsidianVaults  # noqa: F401
 from .indexed_repos import IndexedRepos  # noqa: F401
