+++ tests/test_db_models/test_dl_models.py
@@ -5,15 +5,14 @@ Unit tests for the DLInputs, DLDocuments, and DLChunks SQLAlchemy models.
 """
 
 from datetime import datetime, timedelta, timezone
-from unittest.mock import Mock
 
 import pytest
 
 from wembed_core import AppConfig
 from wembed_core.database import DatabaseService
-from wembed_core.models.dl_chunks import DLChunks
-from wembed_core.models.dl_documents import DLDocuments
-from wembed_core.models.dl_inputs import DLInputs
+from wembed_core.models.dl_doc.dl_chunks import DLChunks
+from wembed_core.models.dl_doc.dl_documents import DLDocuments
+from wembed_core.models.dl_doc.dl_inputs import DLInputs
 
 
 class TestDLModels:
